Refactor repository imports to use server-specific paths
