+++ src/routes/admin/entry/[...slug]/+page.svelte
@@ -4,7 +4,7 @@
 	import type { PageData } from './$types';
 	import { debounce } from '$lib/utils';
 	import { beforeNavigate } from '$app/navigation';
-	import { type LinkPalletData } from '$lib/repository/AdminEntryRepository';
+	import { type LinkPalletData } from '$lib/server/repository/AdminEntryRepository';
 	import LinkPallet from '../../LinkPallet.svelte';
 	import { onMount } from 'svelte';
