+++ src/routes/api/entry/by-title/[title]/+server.ts
@@ -1,5 +1,5 @@
 import type { RequestHandler } from '@sveltejs/kit';
-import { PublicEntryRepository } from '$lib/repository/PublicEntryRepository';
+import { PublicEntryRepository } from '$lib/server/repository/PublicEntryRepository';
 
 export const GET: RequestHandler = async ({ params }) => {
 	const title = params.title;
