+++ src/routes/feed/+server.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from '@sveltejs/kit';
 import { create } from 'xmlbuilder2';
-import { PublicEntryRepository } from '$lib/repository/PublicEntryRepository';
+import { PublicEntryRepository } from '$lib/server/repository/PublicEntryRepository';
 import { renderHTMLByEntry } from '$lib/markdown';
 import { convert } from 'html-to-text';
